+++ apis/features/v1/groupversion_info.go
@@ -17,7 +17,7 @@ limitations under the License.
 // +kubebuilder:object:generate=true
 // +groupName=features.opendatahub.io
 
-// Package v1 contains API Schema definitions for the datasciencecluster v1 API group
+// Package v1 contains API Schema definitions for the feature v1 API group
 package v1
 
 import (
