+++ pkg/cluster/const.go
@@ -1,23 +1,12 @@
 package cluster
 
-import "k8s.io/apimachinery/pkg/runtime/schema"
-
-var (
-	KnativeServingGVK = schema.GroupVersionKind{
-		Group:   "operator.knative.dev",
-		Version: "v1beta1",
-		Kind:    "KnativeServing",
-	}
-
-	OpenshiftIngressGVK = schema.GroupVersionKind{
-		Group:   "config.openshift.io",
-		Version: "v1",
-		Kind:    "Ingress",
-	}
-
-	ServiceMeshControlPlaneGVK = schema.GroupVersionKind{
-		Group:   "maistra.io",
-		Version: "v2",
-		Kind:    "ServiceMeshControlPlane",
-	}
+const (
+	// ManagedRhods defines expected addon catalogsource.
+	ManagedRhods Platform = "addon-managed-odh-catalog"
+	// SelfManagedRhods defines display name in csv.
+	SelfManagedRhods Platform = "Red Hat OpenShift Data Science"
+	// OpenDataHub defines display name in csv.
+	OpenDataHub Platform = "Open Data Hub Operator"
+	// Unknown indicates that operator is not deployed using OLM.
+	Unknown Platform = ""
 )
