+++ pkg/cluster/doc.go
@@ -1,0 +1,7 @@
+// Package cluster contains utility functions used to operate on cluster resources.
+//
+// This includes getting cluster domain, operator namespace and CSV,
+// defining different GVK being used in the project,
+// config metadata on reousrces,
+// managing reousrces like rolebinding, secret and configmap.
+package cluster
