+++ pkg/cluster/gvk/gvk.go
@@ -1,0 +1,37 @@
+package gvk
+
+import "k8s.io/apimachinery/pkg/runtime/schema"
+
+var (
+	KnativeServing = schema.GroupVersionKind{
+		Group:   "operator.knative.dev",
+		Version: "v1beta1",
+		Kind:    "KnativeServing",
+	}
+
+	OpenshiftIngress = schema.GroupVersionKind{
+		Group:   "config.openshift.io",
+		Version: "v1",
+		Kind:    "Ingress",
+	}
+
+	ServiceMeshControlPlane = schema.GroupVersionKind{
+		Group:   "maistra.io",
+		Version: "v2",
+		Kind:    "ServiceMeshControlPlane",
+	}
+
+	OdhApplication = schema.GroupVersionKind{
+		Group:   "dashboard.opendatahub.io",
+		Version: "v1",
+		Kind:    "OdhApplication",
+	}
+	OdhDocument = schema.GroupVersionKind{
+		Group:   "dashboard.opendatahub.io",
+		Version: "v1", Kind: "OdhDocument",
+	}
+	OdhQuickStart = schema.GroupVersionKind{
+		Group:   "console.openshift.io",
+		Version: "v1", Kind: "OdhQuickStart",
+	}
+)
